+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.aegir.wallet",
+    package = "com.aegirwallet",
 )
 
 android_resource(
     name = "res",
-    package = "com.aegir.wallet",
+    package = "com.aegirwallet",
     res = "src/main/res",
 )
